pyproject.toml
[project]
name = "technical-test-api"
version = "0.1.0"
description = "Technical test template - FastAPI with PostgreSQL and Redis"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "sqlmodel==0.0.22",
    "asyncpg==0.29.0",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "redis==5.1.1",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "coverage==7.6.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
